+++ src/main/java/nl/tudelft/skills/repository/research/ResearchConsentRepository.java
@@ -1,0 +1,34 @@
+/*
+ * Skill Circuits
+ * Copyright (C) 2025 - Delft University of Technology
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+package nl.tudelft.skills.repository.research;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+
+import nl.tudelft.skills.model.research.ResearchConsent;
+
+public interface ResearchConsentRepository
+		extends JpaRepository<ResearchConsent, ResearchConsent.ResearchConsentId> {
+
+	@Modifying
+	@Query("delete from ResearchConsent reseachConsent where reseachConsent.id.researchInfoId = :#{#researchInfoId}")
+	void deleteAllByResearchInfoId(@Param("researchInfoId") Long researchInfoId);
+
+}